Moved all domain beans into the beans package. Consolidated all interfaces in the interfaces package. Renamed CompanyDomain to Company.
