+++ src/main/java/ch/zhaw/psit4/domain/ConfigWriter.java
@@ -1,10 +1,10 @@
 package ch.zhaw.psit4.domain;
 
-import ch.zhaw.psit4.domain.dialplan.DialPlanContext;
+import ch.zhaw.psit4.domain.beans.DialPlanContext;
+import ch.zhaw.psit4.domain.beans.SipClient;
 import ch.zhaw.psit4.domain.exceptions.InvalidConfigurationException;
 import ch.zhaw.psit4.domain.interfaces.DialPlanConfigurationInterface;
 import ch.zhaw.psit4.domain.interfaces.SipClientConfigurationInterface;
-import ch.zhaw.psit4.domain.sipclient.SipClient;
 
 import java.util.List;
