+++ src/main/java/ch/zhaw/psit4/domain/beans/DialPlanContext.java
@@ -1,4 +1,4 @@
-package ch.zhaw.psit4.domain.dialplan;
+package ch.zhaw.psit4.domain.beans;
 
 import java.util.List;
