+++ src/main/java/ch/zhaw/psit4/domain/dialplan/applications/DialApp.java
@@ -1,7 +1,7 @@
 package ch.zhaw.psit4.domain.dialplan.applications;
 
-import ch.zhaw.psit4.domain.dialplan.interfaces.DialPlanAppInterface;
-import ch.zhaw.psit4.domain.sipclient.SipClient;
+import ch.zhaw.psit4.domain.beans.SipClient;
+import ch.zhaw.psit4.domain.interfaces.DialPlanAppInterface;
 
 import java.util.List;
