+++ src/main/java/ch/zhaw/psit4/domain/sipclient/SipClientConfigurationChanSip.java
@@ -1,5 +1,6 @@
 package ch.zhaw.psit4.domain.sipclient;
 
+import ch.zhaw.psit4.domain.beans.SipClient;
 import ch.zhaw.psit4.domain.helper.SipClientValidator;
 import ch.zhaw.psit4.domain.interfaces.SipClientConfigurationInterface;
