+++ src/test/java/ch/zhaw/psit4/domain/helper/DialPlanContextValidatorTest.java
@@ -1,7 +1,7 @@
 package ch.zhaw.psit4.domain.helper;
 
-import ch.zhaw.psit4.domain.dialplan.DialPlanContext;
-import ch.zhaw.psit4.domain.dialplan.DialPlanExtension;
+import ch.zhaw.psit4.domain.beans.DialPlanContext;
+import ch.zhaw.psit4.domain.beans.DialPlanExtension;
 import ch.zhaw.psit4.domain.exceptions.InvalidConfigurationException;
 import ch.zhaw.psit4.testsupport.fixtures.domain.DialPlanGenerator;
 import org.junit.Before;
