+++ src/test/java/ch/zhaw/psit4/domain/helper/SipClientValidatorTest.java
@@ -1,7 +1,7 @@
 package ch.zhaw.psit4.domain.helper;
 
+import ch.zhaw.psit4.domain.beans.SipClient;
 import ch.zhaw.psit4.domain.exceptions.InvalidConfigurationException;
-import ch.zhaw.psit4.domain.sipclient.SipClient;
 import ch.zhaw.psit4.testsupport.fixtures.domain.SipClientGenerator;
 import org.junit.Before;
 import org.junit.Test;
