+++ src/test/java/ch/zhaw/psit4/domain/sipclient/SipClientConfigurationChanSipTest.java
@@ -1,5 +1,6 @@
 package ch.zhaw.psit4.domain.sipclient;
 
+import ch.zhaw.psit4.domain.beans.SipClient;
 import ch.zhaw.psit4.domain.exceptions.InvalidConfigurationException;
 import ch.zhaw.psit4.testsupport.convenience.InputStreamStringyfier;
 import ch.zhaw.psit4.testsupport.fixtures.domain.SipClientGenerator;
